refactor: declare and reuse MUTATION_RANGE_REGEX
